requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
atomicwrites==1.4.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
packaging==25.0
playwright==1.54.0
pluggy==1.6.0
py==1.11.0
pyee==13.0.0
Pygments==2.19.2
pytest==7.4.4
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-slugify==8.0.4
requests==2.32.4
six==1.17.0
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.14.0
urllib3==2.5.0
